[FIX] mail: inbox in mobile messaging menu
Previous PR adds "Inbox" dedicated navbar in mobile discuss app,
distinct from "Notifications" (previously named "All").

This was working in the discuss app but not messaging menu. This
is caused by some spaghetti code in which the mailbox content part
was discuss app template that is not available in messaging menu.

This commit fixes the issue by retrieving the "DiscussContent" part
into a dedicated component named `DiscussContent`, which is used in
both discuss app and messaging menu. With this change, the discuss
app in mobile uses fully the messaging menu and the messaging menu
can display mailboxes in a `<Thread>`.

Part of task-4967066
