+++ addons/im_livechat/static/src/core/web/discuss_content_patch.js
@@ -1,7 +1,7 @@
-import { Discuss } from "@mail/core/public_web/discuss";
+import { DiscussContent } from "@mail/core/public_web/discuss_content";
 import { patch } from "@web/core/utils/patch";
 
-patch(Discuss.prototype, {
+patch(DiscussContent.prototype, {
     actionPanelAutoOpenFn() {
         if (!this.threadActions.activeAction) {
             this.threadActions.actions.find((a) => a.id === "livechat-info")?.open();
