+++ addons/mail/static/src/core/public_web/discuss.js
@@ -1,33 +1,17 @@
-import { AutoresizeInput } from "@mail/core/common/autoresize_input";
-import { Composer } from "@mail/core/common/composer";
-import { CountryFlag } from "@mail/core/common/country_flag";
-import { ActionList } from "@mail/core/common/action_list";
-import { ImStatus } from "@mail/core/common/im_status";
-import { Thread } from "@mail/core/common/thread";
-import { useThreadActions } from "@mail/core/common/thread_actions";
-import { ThreadIcon } from "@mail/core/common/thread_icon";
 import { DiscussSidebar } from "@mail/core/public_web/discuss_sidebar";
 import { useMessageScrolling } from "@mail/utils/common/hooks";
 
-import { Component, useRef, useState, useExternalListener, useEffect, useSubEnv } from "@odoo/owl";
+import { Component, useRef, useExternalListener, useEffect, useSubEnv } from "@odoo/owl";
 import { getActiveHotkey } from "@web/core/hotkeys/hotkey_service";
 
-import { _t } from "@web/core/l10n/translation";
 import { useService } from "@web/core/utils/hooks";
-import { FileUploader } from "@web/views/fields/file_handler";
+import { DiscussContent } from "@mail/core/public_web/discuss_content";
 import { MessagingMenu } from "@mail/core/public_web/messaging_menu";
 
 export class Discuss extends Component {
     static components = {
-        ActionList,
-        AutoresizeInput,
-        CountryFlag,
+        DiscussContent,
         DiscussSidebar,
-        Thread,
-        ThreadIcon,
-        Composer,
-        FileUploader,
-        ImStatus,
         MessagingMenu,
     };
     static props = {
@@ -41,9 +25,7 @@ export class Discuss extends Component {
         super.setup();
         this.store = useService("mail.store");
         this.messageHighlight = useMessageScrolling();
-        this.contentRef = useRef("content");
         this.root = useRef("root");
-        this.state = useState({ jumpThreadPresent: 0 });
         this.orm = useService("orm");
         this.effect = useService("effect");
         this.ui = useService("ui");
@@ -51,8 +33,6 @@ export class Discuss extends Component {
             inDiscussApp: true,
             messageHighlight: this.messageHighlight,
         });
-        this.notification = useService("notification");
-        this.threadActions = useThreadActions();
         useExternalListener(
             window,
             "keydown",
@@ -87,55 +67,9 @@ export class Discuss extends Component {
                 () => [this.thread, this.ui.isSmall]
             );
         }
-        useEffect(
-            () => this.actionPanelAutoOpenFn(),
-            () => [this.thread]
-        );
-    }
-
-    actionPanelAutoOpenFn() {
-        const memberListAction = this.threadActions.actions.find((a) => a.id === "member-list");
-        if (!memberListAction) {
-            return;
-        }
-        if (this.store.discuss.isMemberPanelOpenByDefault) {
-            if (!this.threadActions.activeAction) {
-                memberListAction.open();
-            } else if (this.threadActions.activeAction === memberListAction) {
-                return; // no-op (already open)
-            } else {
-                this.store.discuss.isMemberPanelOpenByDefault = false;
-            }
-        }
     }
 
     get thread() {
         return this.props.thread || this.store.discuss.thread;
-    }
-
-    async onFileUploaded(file) {
-        await this.thread.notifyAvatarToServer(file.data);
-        this.notification.add(_t("The avatar has been updated!"), { type: "success" });
-    }
-
-    async renameThread(name) {
-        await this.thread.rename(name);
-    }
-
-    async updateThreadDescription(description) {
-        const newDescription = description.trim();
-        if (!newDescription && !this.thread.description) {
-            return;
-        }
-        if (newDescription !== this.thread.description) {
-            await this.thread.notifyDescriptionToServer(newDescription);
-        }
-    }
-
-    async renameGuest(name) {
-        const newName = name.trim();
-        if (this.store.self.name !== newName) {
-            await this.store.self.updateGuestName(newName);
-        }
     }
 }
