+++ addons/mail/static/src/core/public_web/messaging_menu.js
@@ -12,9 +12,10 @@ import { _t } from "@web/core/l10n/translation";
 import { registry } from "@web/core/registry";
 import { useService } from "@web/core/utils/hooks";
 import { getActiveHotkey } from "@web/core/hotkeys/hotkey_service";
+import { DiscussContent } from "./discuss_content";
 
 export class MessagingMenu extends Component {
-    static components = { CountryFlag, Dropdown, NotificationItem, ImStatus };
+    static components = { CountryFlag, DiscussContent, Dropdown, NotificationItem, ImStatus };
     static props = [];
     static template = "mail.MessagingMenu";
