+++ addons/mail/static/src/discuss/core/public_web/discuss_content_patch.js
@@ -1,13 +1,13 @@
-import { Discuss } from "@mail/core/public_web/discuss";
+import { DiscussContent } from "@mail/core/public_web/discuss_content";
 import { Call } from "@mail/discuss/call/common/call";
 import { PipBanner } from "@mail/discuss/call/common/pip_banner";
 import { useService } from "@web/core/utils/hooks";
 
 import { patch } from "@web/core/utils/patch";
 
-Object.assign(Discuss.components, { Call, PipBanner });
+Object.assign(DiscussContent.components, { Call, PipBanner });
 
-patch(Discuss.prototype, {
+patch(DiscussContent.prototype, {
     setup() {
         super.setup(...arguments);
         this.rtc = useService("discuss.rtc");
